# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Rest - CRUD Usuarios y Artículos",
    "description": "API REST con autenticación JWT, gestión de usuarios y artículos con almacenamiento en JSON",
    "contact": {
      "email": "admin@example.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Servidor Local"
    }
  ],
  "paths": {
...
    "/api/articles": {
      "get": {
        "summary": "Obtener lista de artículos",
        "tags": ["Artículos"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de artículos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "nombre": {
                        "type": "string"
                      },
                      "precio": {
                        "type": "number",
                        "format": "float"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear nuevo artículo",
        "tags": ["Artículos"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nombre", "precio"],
                "properties": {
                  "nombre": {
                    "type": "string",
                    "example": "Laptop"
                  },
                  "precio": {
                    "type": "number",
                    "format": "float",
                    "example": 999.99
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Artículo creado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "article": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validación fallida"
          }
        }
      }
    },
    "/api/articles/{id}": {
      "get": {
        "summary": "Obtener artículo por ID",
        "tags": ["Artículos"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del artículo",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del artículo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "nombre": {
                      "type": "string"
                    },
                    "precio": {
                      "type": "number",
                      "format": "float"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Artículo no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar artículo",
        "tags": ["Artículos"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del artículo",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "precio": {
                    "type": "number",
                    "format": "float"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Artículo actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "article": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Artículo no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar artículo",
        "tags": ["Artículos"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del artículo",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artículo eliminado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Artículo no encontrado"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Login con usuario y contraseña para obtener el token JWT"
      }
    }
  }
}